refactor: move json serialization helpers into client lib, use in both tauri plugins
